Introducing FieldSet and RowField
This two new field types are containers, and not fields on it's own. A
fieldset is expected to be composed of Rows, and a Row of different fields,
cascading rows is not honored, and actually makes no sense.

There are helpers method to tell if field is a field set.

RowField is capable of calculating necessary classes to expand row content.

Fieldset with reference and reference_many fields row are resolved